change to doubles

+++ costas_loop.h
@@ -14,36 +14,36 @@ extern "C"
 #include <stdbool.h>
 #include <complex.h>
 
-void create_costas(float);
-void create_control_loop(float, float, float);
+void create_costas(double);
+void create_control_loop(double, double, double);
 void update_gains(void);
-void advance_loop(float);
+void advance_loop(double);
 void phase_wrap(void);
 void frequency_limit(void);
-float phase_detector(complex float);
+double phase_detector(complex double);
 
 // Setters
 
-void set_loop_bandwidth(float);
-void set_damping_factor(float);
-void set_alpha(float);
-void set_beta(float);
-void set_frequency(float);
-void set_phase(float);
-void set_max_freq(float);
-void set_min_freq(float);
+void set_loop_bandwidth(double);
+void set_damping_factor(double);
+void set_alpha(double);
+void set_beta(double);
+void set_frequency(double);
+void set_phase(double);
+void set_max_freq(double);
+void set_min_freq(double);
 void set_costas_enable(bool);
 
 // Getters
 
-float get_loop_bandwidth(void);
-float get_damping_factor(void);
-float get_alpha(void);
-float get_beta(void);
-float get_frequency(void);
-float get_phase(void);
-float get_max_freq(void);
-float get_min_freq(void);
+double get_loop_bandwidth(void);
+double get_damping_factor(void);
+double get_alpha(void);
+double get_beta(void);
+double get_frequency(void);
+double get_phase(void);
+double get_max_freq(void);
+double get_min_freq(void);
 bool get_costas_enable(void);
 
 #ifdef __cplusplus
